End Of Section 11

+++ src/SearchService/Consumers/AuctionFinishedConsumer.cs
@@ -13,7 +13,7 @@ public class AuctionFinishedConsumer : IConsumer<AuctionFinished>
         if (context.Message.ItemSold)
         {
             auction.Winner = context.Message.Winner;
-            auction.SoldAmount = context.Message.Amount;
+            auction.SoldAmount = (int)context.Message.Amount;
         }
         auction.Status = "Finished";
         await auction.SaveAsync();
